refactor: changes to maintain screens seperately
