+++ screens/GameScreen.tsx
@@ -38,10 +38,10 @@ function GameScreen() {
       <LinearGradient colors={["#00D2FF","#3A7BD5"]}>
         <View style={styles.rootContainer}>
           <Text style={styles.title}>X's Turn</Text>
-          <Image style={[styles.gameTile,{top:320,left:30}]} source={require("./images/StateX.png")}></Image>
+          <Image style={[styles.gameTile,{top:320,left:30}]} source={require("../images/StateX.png")}></Image>
           <Pressable onPress={handlePress}>
           <View style={styles.gameSquare}>
-            <Image style={{height:340,width:340,marginTop:5,marginLeft:5}} source={require("./images/play_board.png")}/>
+            <Image style={{height:340,width:340,marginTop:5,marginLeft:5}} source={require("../images/play_board.png")}/>
           </View>
           </Pressable>
         </View>
